inference_3d/fin_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.3
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
antlr4-python3-runtime=4.9.3=pypi_0
attrs=25.3.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.8.3=hbd8a1cb_0
certifi=2025.8.3=pypi_0
charset-normalizer=3.4.3=pypi_0
coloredlogs=15.0.1=pypi_0
cuda-cccl=12.1.55=0
cuda-command-line-tools=12.1.0=0
cuda-compiler=12.1.0=0
cuda-cudart=12.1.55=0
cuda-cudart-dev=12.1.55=0
cuda-cudart-static=12.1.55=0
cuda-cuobjdump=12.1.55=0
cuda-cupti=12.1.62=0
cuda-cupti-static=12.1.62=0
cuda-cuxxfilt=12.1.55=0
cuda-documentation=12.1.55=0
cuda-driver-dev=12.1.55=0
cuda-gdb=12.1.55=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-libraries-static=12.1.0=0
cuda-nsight=12.1.55=0
cuda-nsight-compute=12.1.0=0
cuda-nvcc=12.1.66=0
cuda-nvdisasm=12.1.55=0
cuda-nvml-dev=12.1.55=0
cuda-nvprof=12.1.55=0
cuda-nvprune=12.1.55=0
cuda-nvrtc=12.1.55=0
cuda-nvrtc-dev=12.1.55=0
cuda-nvrtc-static=12.1.55=0
cuda-nvtx=12.1.66=0
cuda-nvvp=12.1.55=0
cuda-opencl=12.1.56=0
cuda-opencl-dev=12.1.56=0
cuda-profiler-api=12.1.55=0
cuda-sanitizer-api=12.1.55=0
cuda-toolkit=12.1.0=0
cuda-tools=12.1.0=0
cuda-visual-tools=12.1.0=0
diffusers=0.34.0=pypi_0
einops=0.8.1=pypi_0
expat=2.7.1=h6a678d5_0
filelock=3.19.1=pypi_0
flatbuffers=25.2.10=pypi_0
fsspec=2025.7.0=pypi_0
gds-tools=1.6.0.25=0
hf-xet=1.1.8=pypi_0
huggingface-hub=0.34.4=pypi_0
humanfriendly=10.0=pypi_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
importlib-metadata=8.7.0=pypi_0
jinja2=3.1.6=pypi_0
jsonschema=4.25.1=pypi_0
jsonschema-specifications=2025.4.1=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcublas-static=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcufft-static=11.0.2.4=0
libcufile=1.6.0.25=0
libcufile-dev=1.6.0.25=0
libcufile-static=1.6.0.25=0
libcurand=10.3.2.56=0
libcurand-dev=10.3.2.56=0
libcurand-static=10.3.2.56=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusolver-static=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libcusparse-static=12.0.2.55=0
libffi=3.4.4=h6a678d5_1
libgcc=15.1.0=h767d61c_4
libgcc-ng=15.1.0=h69a702a_4
libgomp=15.1.0=h767d61c_4
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnpp-static=12.0.2.50=0
libnvjitlink=12.1.55=0
libnvjitlink-dev=12.1.55=0
libnvjpeg=12.1.0.39=0
libnvjpeg-dev=12.1.0.39=0
libnvjpeg-static=12.1.0.39=0
libnvvm-samples=12.1.55=0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
llvmlite=0.44.0=pypi_0
markupsafe=3.0.2=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.5=h7934f7d_0
networkx=3.4.2=pypi_0
nsight-compute=2023.1.0.15=0
numba=0.61.2=pypi_0
numpy=2.2.6=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.9.86=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
omegaconf=2.3.0=pypi_0
onnxruntime=1.22.1=pypi_0
opencv-python-headless=4.12.0.88=pypi_0
openssl=3.5.2=h26f9b46_0
packaging=25.0=pypi_0
pillow=11.0.0=pypi_0
pip=25.2=pyh8b19718_0
platformdirs=4.3.8=pypi_0
pooch=1.8.2=pypi_0
protobuf=6.32.0=pypi_0
pthread-stubs=0.3=h0ce48e5_1
pymatting=1.1.14=pypi_0
python=3.10.18=h1a3bd86_0
pyyaml=6.0.2=pypi_0
readline=8.3=hc2a1206_0
referencing=0.36.2=pypi_0
regex=2025.7.34=pypi_0
rembg=2.0.67=pypi_0
requests=2.32.5=pypi_0
rpds-py=0.27.0=pypi_0
safetensors=0.6.2=pypi_0
scikit-image=0.25.2=pypi_0
scipy=1.15.3=pypi_0
setuptools=80.9.0=pyhff2d567_0
sqlite=3.50.2=hb25bd0a_1
sympy=1.13.1=pypi_0
tifffile=2025.5.10=pypi_0
tk=8.6.15=h54e0aa7_0
tokenizers=0.21.4=pypi_0
torch=2.5.1+cu121=pypi_0
torchaudio=2.5.1+cu121=pypi_0
torchmcubes=0.1.0=pypi_0
torchvision=0.20.1+cu121=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.55.2=pypi_0
trimesh=4.7.4=pypi_0
triton=3.1.0=pypi_0
typing-extensions=4.15.0=pypi_0
tzdata=2025b=h04d1e81_0
urllib3=2.5.0=pypi_0
wheel=0.45.1=py310h06a4308_0
xformers=0.0.32.post2=pypi_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
zipp=3.23.0=pypi_0
zlib=1.2.13=h5eee18b_1
